+++ src/WebUI/src/assets/themes/oruga-tailwind/icons/chart.ts
@@ -1,0 +1,11 @@
+export default {
+  prefix: 'crpg',
+  iconName: 'chart',
+  icon: [
+    32,
+    32,
+    [],
+    'e001',
+    'M3 2v28h26v-1.081h-1.083V16.083h-3.25V28.92h-1.084V10.125l-3.25.066V28.92H19.25v-9.044L16 19.86v9.059h-1.083V5.45h-3.25v23.469h-1.084V12.154h-3.25V28.92H6.25V2H3Z',
+  ],
+};
