+++ src/WebUI/src/assets/themes/oruga-tailwind/icons/experience.ts
@@ -6,6 +6,6 @@ export default {
     32,
     [],
     'e001',
-    'M16.168 10.234c-.454.113-.85.17-1.191.17-.326 0-.638-.021-.936-.064-.837.596-1.844 1.567-3.02 2.915 1.715 2.68 3.084 5.084 4.105 7.211H9.829a43.635 43.635 0 0 0-1.745-3.382c-1.574 2.24-2.787 4.226-3.638 5.957-.283.07-.702.163-1.255.276L2 23.594c1.191-2.128 2.971-4.709 5.34-7.744-1.461-2.283-3.113-4.29-4.957-6.02h6.616c.326.694.78 1.489 1.362 2.382.113-.127.333-.383.66-.766.325-.397.58-.695.765-.893.184-.199.425-.44.723-.723.582-.554 1.184-.944 1.809-1.17.24-.071.73-.142 1.467-.213.738-.085 1.27-.234 1.596-.447-.34.78-.745 1.525-1.213 2.234ZM23 17.807c0 .809.036 1.695.107 2.66h-4.68c-.581-2.256-.872-4.652-.872-7.191 0-1.29.078-2.44.234-3.446h11.53c.412 1.602.639 3.51.681 5.722-2.099.936-4.432 1.546-7 1.83v.425Zm.022-1.893c.908-.114 1.667-.34 2.277-.68a7.85 7.85 0 0 0-1.83-4 39.788 39.788 0 0 0-.447 4.68Z',
+    'M16.168 12.234c-.454.113-.85.17-1.191.17-.326 0-.638-.021-.936-.064-.837.596-1.844 1.567-3.02 2.915 1.715 2.68 3.084 5.084 4.105 7.211H9.829a43.635 43.635 0 0 0-1.745-3.382c-1.574 2.24-2.787 4.226-3.638 5.957-.283.07-.702.163-1.255.276L2 25.594c1.191-2.128 2.971-4.709 5.34-7.744-1.461-2.283-3.113-4.29-4.957-6.02h6.616c.326.694.78 1.489 1.362 2.382.113-.127.333-.383.66-.766.325-.397.58-.695.765-.893.184-.199.425-.44.723-.724.582-.553 1.184-.943 1.809-1.17.24-.07.73-.141 1.467-.212.738-.085 1.27-.234 1.596-.447-.34.78-.745 1.525-1.213 2.234ZM23 19.807c0 .809.036 1.695.107 2.66h-4.68c-.581-2.256-.872-4.652-.872-7.191 0-1.29.078-2.44.234-3.447h11.53c.412 1.603.639 3.51.681 5.723-2.099.936-4.432 1.546-7 1.83v.425Zm.022-1.893c.908-.114 1.667-.34 2.277-.68a7.85 7.85 0 0 0-1.83-4 39.788 39.788 0 0 0-.447 4.68Z',
   ],
 };
