+++ src/WebUI/src/components/item/ItemDetailGroup.vue
@@ -14,17 +14,16 @@ const { openedItems, closeItemDetail, computeDetailCardYPosition } = useItemDeta
         x: oi.bound.x + oi.bound.width + 8,
         y: computeDetailCardYPosition(oi.bound.y),
       }"
-      class="fixed z-50 cursor-move select-none overflow-hidden rounded-lg bg-base-300 shadow-lg active:ring-1"
+      class="fixed z-50 cursor-move select-none rounded-lg bg-base-300 shadow-lg active:ring-1"
     >
       <OButton
-        class="!absolute right-2 top-2 z-10 cursor-pointer"
+        class="!absolute -right-3 -top-3 z-10 cursor-pointer"
         iconRight="close"
         rounded
-        size="2xs"
+        size="xs"
         variant="secondary"
         @click="closeItemDetail(oi)"
       />
-
       <slot v-bind="oi" />
     </Draggable>
   </Teleport>
