+++ src/WebUI/src/models/timeseries.ts
@@ -1,0 +1,4 @@
+export interface TimeSeries {
+  name: string;
+  data: [Date, number][];
+}
