+++ src/WebUI/src/types/vite-components.d.ts
@@ -9,6 +9,7 @@ declare module 'vue' {
   export interface GlobalComponents {
     ActivityLogItem: typeof import('./../components/moderator/ActivityLogItem.vue')['default']
     Bg: typeof import('./../components/app/Bg.vue')['default']
+    CharacterCharts: typeof import('./../components/character/CharacterCharts.vue')['default']
     CharacterCreateModal: typeof import('./../components/character/CharacterCreateModal.vue')['default']
     CharacterEditForm: typeof import('./../components/character/CharacterEditForm.vue')['default']
     CharacterInventoryDoll: typeof import('./../components/character/inventory/CharacterInventoryDoll.vue')['default']
